Splitting framework from service
The project is composed of several pieces as known, framework is both shared
by the service and by end user apps.

While service is the server it self the one that talks to BlueZ and does all
the magic, I'm investigating the posibility to make this an apk installable
into /system/apk which provides a new permission that sets the gid to
bluetooth as other rules do as well. If this works then the service could be
started by matching broadcasted signals from android.bluetooth