Restructuring code back to previous state, adding a modified Hexdump
Code needed structure changes again, I can't get to both be able to open a
DBus socket from the APK and to register my self as well with SystemServer
so I took the determination that the DBus layer will have to be lauched by
some black magic like modifying the dbus daemon by adding a script in the
middle that starts the required piece.

Also looks like toHex from Android isn't compatible 100% with the one that
java-dbus was using, thus failing during the autentication part.