Synchronizate with GattToolWrapper not with BluetoothGatt
Having so many GattToolWrapper callback methods marked as synchronized was a
serious mistake, as this was preventing the callback to actually happen.

Marking the binder methods as synchronized shouldn't be an issue as we're not
supposed to handle many commands at the same time.

synchronized methods needs to be replaced with ConnectionWrapper.mGattTool ins-
tead.